+++ src/models/movie.interface.ts
@@ -1,0 +1,13 @@
+export interface Movie {
+  id: string;
+  imageUrl: string;
+  name: string;
+  releaseYear: number;
+  genres?: string[];
+}
+
+export interface MovieDetailsData extends Movie {
+  rating: number;
+  duration: number | string;
+  description: string;
+}
